Bump version & rename file
