+++ Renegade/Utils.h
@@ -21,7 +21,7 @@ using std::endl;
 using std::get;
 using Clock = std::chrono::high_resolution_clock;
 
-constexpr std::string_view Version = "dev 1.1.95";
+constexpr std::string_view Version = "dev 1.1.96";
 
 // Evaluation helpers -----------------------------------------------------------------------------
